+++ contracts/access_control/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "access_control"
-version = "0.1.0"
+version = "1.0.0"
 authors = ["Cardinal Cryptography"]
 edition = "2021"
 publish = false
